Fix performance issues for addition
Addition was slow because a new vector of roots was created for every
leaf that is added.
This is fixed by storing the set of roots as a member of the
accumulator and maintaining it while the accumulator is modified.

+++ state.cpp
@@ -343,7 +343,7 @@ uint8_t ForestState::detectRow(uint64_t pos) const
     return row;
 }
 
-uint8_t ForestState::rowOffset(uint64_t pos) const
+uint64_t ForestState::rowOffset(uint64_t pos) const
 {
     uint8_t row = this->detectRow(pos);
     uint64_t marker = this->maxNodes();
